+++ src/main/java/drapomods/settlerbubbles/SettlerBubblesMod.java
@@ -10,6 +10,10 @@ import necesse.engine.registries.PacketRegistry;
 
 @ModEntry
 public class SettlerBubblesMod {
+    public static final String MOD_ID = "drapomods.settlerbubbles";
+    public static final String VERSION = "1.2.0";
+    public static final String GAME_VERSION = "1.3.2";
+
     public ModSettings initSettings() {
         return new SettlerBubblesSettings();
     }